+++ src/components/Header/Header.module.scss
@@ -5,12 +5,6 @@
 	align-items: center;
 }
 
-.profile {
-	display: flex;
-	gap: 20px;
-	align-items: center;
-}
-
 .entry {
 	width: 100px;
 	cursor: pointer;
@@ -28,78 +22,5 @@
 	}
 	&:active {
 		background: #3b29a1;
-	}
-}
-
-.container__btn {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	gap: 20px;
-}
-
-.btn {
-	background-color: #fff;
-	border: none;
-	border-bottom: 1px solid #000;
-	padding-top: 11.5px;
-	padding-bottom: 17.5px;
-	padding-left: 12px;
-	padding-right: 12px;
-	max-height: 100%;
-	width: 100%;
-}
-
-.btns {
-	&-enter {
-		opacity: 0;
-		transform: translateX(50%);
-
-		&-active {
-			transition: all 300ms;
-
-			opacity: 1;
-
-			transform: translateX(0);
-		}
-	}
-
-	&-exit {
-		opacity: 1;
-
-		&-active {
-			transition: all 300ms;
-
-			opacity: 0;
-
-			transform: translateX(50%);
-		}
-	}
-}
-
-.profile:has(.btns-enter-active) {
-	animation: slidein__right 0.7s;
-}
-
-@keyframes slidein__right {
-	0% {
-		transform: translateX(50%);
-	}
-	100% {
-		transform: translateX(-2%);
-	}
-}
-
-.profile:has(.btns-exit-active) {
-	animation: slidein__left 0.7s;
-}
-
-@keyframes slidein__left {
-	0% {
-		transform: translateX(0%);
-	}
-
-	100% {
-		transform: translateX(61%);
 	}
 }
